Hackathon/Lucca Fortino/requirements.txt
# Dependências principais
ultralytics>=8.0.0
torch>=2.0.0
numpy>=1.21.0
opencv-python>=4.8.0
PyQt5>=5.15.0

# Utilitários
pillow>=10.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
psutil>=5.9.0
py-cpuinfo>=9.0.0

# Email
secure-smtplib>=0.1.1
python-dotenv>=1.0.0

# Logging e monitoramento
colorama>=0.4.6